move menu share from main activity to report activity
