+++ app/src/main/res/menu/menu_report.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:context="com.ahmadrosid.dompetku.report.ReportActivity">
+
+    <item
+        android:id="@+id/action_share"
+        android:orderInCategory="100"
+        android:icon="@android:drawable/ic_menu_share"
+        app:showAsAction="always"
+        android:title="Share"/>
+
+</menu>
